llms/agents/agents_dev_space/demos/demo1/pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Answer questions related to Vertex AI documentation using Vertex AI RAG Engine"
authors = [
        { name = "Meltem Subasioglu", email = "msubasioglu@google.com" },
]

license = "Apache License 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tabulate = "^0.9.0"
google-auth = "^2.36.0"
requests = "^2.32.3"
google-adk = "^1.10.0"
google-cloud-aiplatform = {extras = ["evaluation"], version = "^1.110.0"}
llama-index = "^0.12"
frontend = "^0.0.3"
pymupdf = "^1.26.3"
google-generativeai = "^0.8.5"
pillow = "^11.3.0"
google-cloud-modelarmor = "^0.2.8"
flet = "^0.28.3"
flet-desktop = "^0.28.3"
absl-py = "^2.3.1"
pydantic = "^2.11.7"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-mock = "^3.14.0"
scikit-learn = "^1.6.1"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.26.0"
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
